Tying some features together to get a working app
