+++ client/src/App.jsx
@@ -1,14 +1,19 @@
 import './App.css'
 import { Outlet } from 'react-router-dom' 
 import {ApolloClient, InMemoryCache, ApolloProvider} from '@apollo/client'
+
+import Navbar from './components/Navbar';
+
 const client = new ApolloClient({
  uri:'/graphql',
  cache: new InMemoryCache(),
 })
 
 function App() {
+  console.log("App component loaded");
   return (
     <ApolloProvider client = {client}>
+      <Navbar />
       <Outlet/>
     </ApolloProvider>
   )
