+++ client/src/components/navbar.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
-import './App.css'; 
+import '../App.css'; 
 
 const Navbar = () => {
   const [isOpen, setIsOpen] = useState(false);
@@ -14,13 +14,12 @@ const Navbar = () => {
       <div className="navbar-logo">
         <Link to="/">MemorEase</Link>
       </div>
-      
       <div className="navbar-toggle" onClick={toggleMenu}>
         <span className="navbar-toggle-icon"></span>
       </div>
 
       <div className={`navbar-links ${isOpen ? 'open' : ''}`}>
-        <Link to="/Home" onClick={toggleMenu}>Home</Link>
+        <Link to="/" onClick={toggleMenu}>Home</Link>
         <Link to="/Collection" onClick={toggleMenu}>Collection</Link>
         <Link to="/LogOut" onClick={toggleMenu}>Log Out</Link>
       </div>
