+++ client/src/pages/Collection.jsx
@@ -1,0 +1,8 @@
+const Collections = () => {
+    return (
+        <>
+        </>
+    )
+}
+
+export default Collections;
