+++ client/src/pages/Decks.jsx
@@ -1,0 +1,8 @@
+const Decks = () => {
+    return (
+        <>
+        </>
+    )
+}
+
+export default Decks;
