+++ client/src/pages/Signup.jsx
@@ -1,0 +1,8 @@
+const Signup = () => {
+    return (
+        <>
+        </>
+    )
+}
+
+export default Signup;
