+++ server/package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "start": "node server.js",
     "watch": "nodemon",
-    "seed": "node seeders/seed.js"
+    "seed": "node seeds/seed.js"
   },
   "keywords": [],
   "author": "",
